third_party/WebKit/LayoutTests/accessibility/aria-labelledby-overrides-label.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test.js"></script>
</head>
<body id="body">

<label for="target">Label</label>

<div id="a">A</div>
<div id="b">B</div>
<div id="c">C</div>

<input id="target" aria-labelledby="a b c" value="10">

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that if aria-labelledby is used, then label elements are not used");

    if (window.accessibilityController) {
          var axTarget = accessibilityController.accessibleElementById("target");
          shouldBe("axTarget.name", "'A B C'");
          shouldBe("axTarget.nameElementCount()", "3");
    }

</script>

</body>
</html>
